Add base plate to bottom case.
This creates a base plate for the bottom case by filling in the
polygon defined by the walls. The main and thumb areas are done
separately, as this allows us to use a simple 'fill holes' algorithm
that was found on the OpenSCAD mailing list.

This required upgrading the scad-clj library in order to allow
scale options to be passed to extrude-linear.